requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pytelegrambotapi==4.22.0
python-dotenv==1.0.1
requests==2.32.3
telebot==0.0.5
urllib3==2.2.2
